# [Exporter.Geneva] Perf improvement on .NET 8